updated babel.js, gulp 4.x, node 0.12.x, POC gulpfile in es6
