+++ gulpfile-es6.js
@@ -1,0 +1,5 @@
+let gulp = global.gulp;
+
+gulp.task('default', () => {
+        console.log('!hi');
+});
