+++ package.json
@@ -13,9 +13,9 @@
     "url": "http://github.com/cferdinandi/Plugin"
   },
   "devDependencies": {
-    "babel": "^4.4.6",
-    "babel-runtime": "^4.4.6",
-    "babelify": "^5.0.3",
+    "babel": "^5.2.17",
+    "babel-runtime": "^5.2.17",
+    "babelify": "^6.0.2",
     "brfs": "^1.2.0",
     "browserify": "^8.1.0",
     "browserify-ngannotate": "^0.6.0",
@@ -34,10 +34,10 @@
     "exorcist": "^0.1.6",
     "express": "^4.10.7",
     "filter-transform": "^0.1.1",
-    "gulp": "^3.8.11",
+    "gulp": "git://github.com/gulpjs/gulp.git#4.0",
     "gulp-angular-templatecache": "^1.5.0",
     "gulp-autoprefixer": "^0.0.7",
-    "gulp-babel": "^4.0.0",
+    "gulp-babel": "^5.1.0",
     "gulp-bower": "0.0.7",
     "gulp-buffer": "0.0.2",
     "gulp-cached": "^1.0.1",
